figma_test_generator/requirements.txt
streamlit
google-generativeai
Pillow
requests
python-dotenv
playwright